+++ test/lib/studio/profile_sections_test.rb
@@ -201,7 +201,7 @@ class ProfileSectionsTest < Minitest::Test
   end
 
   def test_studio_default_profile_sections_is_the_public_composition_handle
-    assert_equal %i[google name email birthday], Studio.default_profile_sections.map { |s| s[:key] }
+    assert_equal %i[google newsletter name email birthday], Studio.default_profile_sections.map { |s| s[:key] }
   end
 
   # --- the app-capability gate (`if:`) ----------------------------------------